+++ llvm/bindings/go/llvm/bitreader.go
@@ -41,7 +41,7 @@ func ParseBitcodeFile(name string) (Module, error) {
 	defer C.LLVMDisposeMemoryBuffer(buf)
 
 	var m Module
-	if C.LLVMParseBitcode(buf, &m.C, &errmsg) == 0 {
+	if C.LLVMParseBitcode2(buf, &m.C) == 0 {
 		return m, nil
 	}