+++ llvm/tools/llvm-c-test/main.c
@@ -26,6 +26,11 @@ static void print_usage(void) {
   fprintf(stderr, "  * --lazy-module-dump\n");
   fprintf(stderr,
           "    Lazily read bytecode from stdin - print disassembly\n\n");
+  fprintf(stderr, "  * --new-module-dump\n");
+  fprintf(stderr, "    Read bytecode from stdin - print disassembly\n\n");
+  fprintf(stderr, "  * --lazy-new-module-dump\n");
+  fprintf(stderr,
+          "    Lazily read bytecode from stdin - print disassembly\n\n");
   fprintf(stderr, "  * --module-list-functions\n");
   fprintf(stderr,
           "    Read bytecode from stdin - list summary of functions\n\n");
@@ -52,10 +57,14 @@ int main(int argc, char **argv) {
 
   LLVMInitializeCore(pr);
 
-  if (argc == 2 && !strcmp(argv[1], "--lazy-module-dump")) {
-    return module_dump(true);
+  if (argc == 2 && !strcmp(argv[1], "--lazy-new-module-dump")) {
+    return module_dump(true, true);
+  } else if (argc == 2 && !strcmp(argv[1], "--new-module-dump")) {
+    return module_dump(false, true);
+  } else if (argc == 2 && !strcmp(argv[1], "--lazy-module-dump")) {
+    return module_dump(true, false);
   } else if (argc == 2 && !strcmp(argv[1], "--module-dump")) {
-    return module_dump(false);
+    return module_dump(false, false);
   } else if (argc == 2 && !strcmp(argv[1], "--module-list-functions")) {
     return module_list_functions();
   } else if (argc == 2 && !strcmp(argv[1], "--module-list-globals")) {